@@ -352,6 +352,7 @@ func DialContext(ctx context.Context, options ...ClientOptionFunc) (*Client, err
 	if len(c.urls) == 0 {
 		c.urls = []string{DefaultURL}
 	}
+
 	c.urls = canonicalize(c.urls...)
 
 	// If the URLs have auth info, use them here as an alternative to SetBasicAuth
@@ -872,7 +873,7 @@ func (c *Client) sniffer() {
 			c.snifferStop <- true
 			return
 		case <-ticker.C:
-			c.sniff(context.Background(), timeout)
+			_ = c.sniff(context.Background(), timeout)
 		}
 	}
 }
