@@ -53,7 +53,7 @@ func TestIndexGetSettingsURL(t *testing.T) {
 func TestIndexGetSettingsService(t *testing.T) {
 	client := setupTestClientAndCreateIndex(t)
 
-	esversion, err := client.OpensearchVersion(DefaultURL)
+	esversion, err := client.OpensearchVersion("https://opensearch.svc:9200")
 	if err != nil {
 		t.Fatal(err)
 	}
