@@ -82,10 +82,12 @@ func TestGetScript(t *testing.T) {
 	}
 
 	// Cleanup
-	client.PerformRequest(
+	if _, err = client.PerformRequest(
 		context.Background(),
 		PerformRequestOptions{
 			Method: "DELETE",
 			Path:   "/_scripts/" + scriptID,
-		})
+		}); err != nil {
+		t.Fatal(err)
+	}
 }
